+++ KSPAddon.cs
@@ -12,19 +12,18 @@ namespace KRPC.MechJeb {
 
 	[KSPAddon(KSPAddon.Startup.Flight, false)]
 	internal class InitInstance : MonoBehaviour {
-		public void Start() {
-			if(MechJeb.TypesLoaded) {
+		private Vessel activeVessel;
+
+		public void LateUpdate() {
+			if(!MechJeb.TypesLoaded)
+				return;
+
+			// Refresh MechJeb instance when focus changes or a flight is reverted to launch.
+			if(this.activeVessel != FlightGlobals.ActiveVessel) {
+				this.activeVessel = FlightGlobals.ActiveVessel;
 				Logger.Info("Initializing MechJeb instance...");
 				Logger.Info(MechJeb.InitInstance() ? "KRPC.MechJeb is ready!" : "MechJeb found but the instance initialization wasn't successful. Maybe you don't have any MechJeb part attached to the vessel?");
 				MechJeb.ShowErrors();
-			}
-		}
-
-		public void FixedUpdate() {
-			// Needed to fix outdated MechJeb instance when focus changes or a flight is reverted to launch.
-			if(FlightGlobals.ActiveVessel != MechJeb.MasterInstance.vessel) {
-				Logger.Info("MechJeb instance changed, resetting cache...");
-				this.Start();
 			}
 		}
 	}